+++ bin/transporte.dart
@@ -1,0 +1,11 @@
+enum Transporte {
+  //Aula 1 -> Resolvendo o problema dos numeros
+  carro,
+  bicicleta,
+  onibus,
+  aviao,
+  andando,
+  helicoptero,
+  patins,
+  skate,
+}
